requirements.txt
fastapi
uvicorn
streamlit
yfinance
langchain-core
langchain-google-genai
langgraph
PyJWT
passlib
bcrypt==4.0.1
sqlalchemy
matplotlib
requests
python-dotenv